@@ -14,13 +14,14 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../src'))
+
+sys.path.insert(0, os.path.abspath("../src"))
 
 
 # -- Project information -----------------------------------------------------
 import sasctl
 
-project = 'sasctl'
+project = "sasctl"
 copyright = sasctl.__copyright__
 author = sasctl.__author__
 
@@ -40,38 +41,39 @@ release = sasctl.__version__
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.autosummary',
-    'sphinx.ext.viewcode',
-    'sphinx.ext.napoleon',
-    'sphinx.ext.intersphinx',
-    'sphinx.ext.todo'
+    "sphinx.ext.autodoc",
+    "sphinx.ext.autosummary",
+    "sphinx.ext.viewcode",
+    "sphinx.ext.napoleon",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.todo",
 ]
 
-intersphinx_mapping = {'python': ('https://docs.python.org/3', None),
-                       'swat': ('https://sassoftware.github.io/python-swat/', None),
-                       'pytest': ('https://docs.pytest.org/en/latest/', None),
-                       'betamax': ('https://betamax.readthedocs.io/en/latest/', None),
-                       'requests': (
-                           'https://2.python-requests.org/en/master/', None),
-                       'tox': ('https://tox.readthedocs.io/en/latest/', None),
-                       'flake8': ('http://flake8.pycqa.org/en/latest/', None)}
+intersphinx_mapping = {
+    "python": ("https://docs.python.org/3", None),
+    "swat": ("https://sassoftware.github.io/python-swat/", None),
+    "pytest": ("https://docs.pytest.org/en/latest/", None),
+    "betamax": ("https://betamax.readthedocs.io/en/latest/", None),
+    "requests": ("https://2.python-requests.org/en/master/", None),
+    "tox": ("https://tox.readthedocs.io/en/latest/", None),
+    "flake8": ("http://flake8.pycqa.org/en/latest/", None),
+}
 
 todo_include_todos = True
 
 todo_emit_warnings = True
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 #
 # source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
+source_suffix = ".rst"
 
 # The master toctree document.
-master_doc = 'index'
+master_doc = "index"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
@@ -83,10 +85,10 @@ language = None
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path .
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 # The name of the Pygments (syntax highlighting) style to use.
-pygments_style = 'sphinx'
+pygments_style = "sphinx"
 
 
 # -- Options for HTML output -------------------------------------------------
@@ -94,7 +96,7 @@ pygments_style = 'sphinx'
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'alabaster'
+html_theme = "alabaster"
 # html_theme = 'sphinx_rtd_theme'
 
 # Theme options are theme-specific and customize the look and feel of a theme
@@ -102,14 +104,14 @@ html_theme = 'alabaster'
 # documentation.
 #
 html_theme_options = {
-    'description': 'Python package and CLI for user-friendly integration with SAS Viya',
-    'fixed_sidebar': True
+    "description": "Python package and CLI for user-friendly integration with SAS Viya",
+    "fixed_sidebar": True,
 }
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = ["_static"]
 
 # Custom sidebar templates, must be a dictionary that maps document names
 # to template names.
@@ -120,19 +122,13 @@ html_static_path = ['_static']
 # 'searchbox.html']``.
 #
 # html_sidebars = {}
-html_sidebars = {
-    '**': [
-        'about.html',
-        'searchbox.html',
-        'navigation.html'
-    ]
-}
+html_sidebars = {"**": ["about.html", "searchbox.html", "navigation.html"]}
 
 
 # -- Options for HTMLHelp output ---------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'sasctldoc'
+htmlhelp_basename = "sasctldoc"
 
 
 # -- Options for LaTeX output ------------------------------------------------
@@ -141,15 +137,12 @@ latex_elements = {
     # The paper size ('letterpaper' or 'a4paper').
     #
     # 'papersize': 'letterpaper',
-
     # The font size ('10pt', '11pt' or '12pt').
     #
     # 'pointsize': '10pt',
-
     # Additional stuff for the LaTeX preamble.
     #
     # 'preamble': '',
-
     # Latex figure (float) alignment
     #
     # 'figure_align': 'htbp',
@@ -159,8 +152,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'sasctl.tex', 'sasctl Documentation',
-     'SAS', 'manual'),
+    (master_doc, "sasctl.tex", "sasctl Documentation", "SAS", "manual"),
 ]
 
 
@@ -168,10 +160,7 @@ latex_documents = [
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [
-    (master_doc, 'sasctl', 'sasctl Documentation',
-     [author], 1)
-]
+man_pages = [(master_doc, "sasctl", "sasctl Documentation", [author], 1)]
 
 
 # -- Options for Texinfo output ----------------------------------------------
@@ -180,9 +169,15 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'sasctl', 'sasctl Documentation',
-     author, 'sasctl', 'One line description of project.',
-     'Miscellaneous'),
+    (
+        master_doc,
+        "sasctl",
+        "sasctl Documentation",
+        author,
+        "sasctl",
+        "One line description of project.",
+        "Miscellaneous",
+    ),
 ]
 
 
