@@ -4,7 +4,7 @@
 # Copyright © 2019, SAS Institute Inc., Cary, NC, USA.  All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-__version__ = "1.6.4"
+__version__ = "1.8.1"
 __author__ = "SAS"
 __credits__ = [
     "Yi Jian Ching",
@@ -17,7 +17,8 @@ __credits__ = [
 ]
 __license__ = "Apache 2.0"
 __copyright__ = (
-    "Copyright © 2019, SAS Institute Inc., " "Cary, NC, USA.  All Rights Reserved."
+    "Copyright © 2019, SAS Institute Inc., ",
+    "Cary, NC, USA.  All Rights Reserved.",
 )
 
 import logging
