@@ -6,7 +6,8 @@
 
 
 from .service import Service
-from ..core import uri_as_str
+from ..core import current_session, uri_as_str
+
 
 
 class SentimentAnalysis(Service):
@@ -71,6 +72,11 @@ class SentimentAnalysis(Service):
         :meth:`cas_management.get_table <.CASManagement.get_table>`
 
         """
+        if current_session().version_info() >= 4:
+            raise RuntimeError(
+                "The Sentiment Analysis service was removed from Viya 4."
+            )
+
         if documents is None:
             raise TypeError("`documents` cannot be None.")
 
