@@ -6,7 +6,7 @@
 
 
 from .service import Service
-from sasctl.core import uri_as_str
+from sasctl.core import current_session, uri_as_str
 
 
 class TextParsing(Service):
@@ -100,6 +100,8 @@ class TextParsing(Service):
         :meth:`cas_management.get_table <.CASManagement.get_table>`
 
         """
+        if current_session().version_info() >= 4:
+            raise RuntimeError("The Text Parsing service was removed from Viya 4.")
 
         if documents is None:
             raise TypeError("`documents` cannot be None.")
