@@ -12,7 +12,6 @@ import netrc
 import os
 import re
 import ssl
-import sys
 import warnings
 from datetime import datetime, timedelta
 from uuid import UUID, uuid4
@@ -98,7 +97,7 @@ def _filter_token(r):
     if hasattr(r, "headers") and "X-Consul-Token" in r.headers:
         r.headers["X-Consul-Token"] = "[redacted]"
 
-    # Redact "access_token":"<token>" in response from Logon service
+    # Redact "access_token":"<token>" in response from SASLogon service
     if hasattr(r, "_content"):
         r._content = _redact('(?<=access_token":")[^"]*', "[redacted]", r._content)
 
@@ -164,7 +163,7 @@ class OAuth2Token(requests.auth.AuthBase):
         refresh_token=None,
         expiration=None,
         expires_in=None,
-        **kwargs
+        **kwargs,
     ):
         self.access_token = access_token
         self.refresh_token = refresh_token
@@ -260,6 +259,13 @@ class Session(requests.Session):
         to true.  Ignored for HTTP connections.
     token : str, optional
         OAuth token to use for authorization.
+    client_id : str, optional
+        Client ID requesting access.  Use if connection to Viya should be
+        made using `client_credentials` method.
+    client_secret : str, optional
+        Client secret for client requesting access. Required if `client_id`
+        is provided.
+
 
     Attributes
     ----------
@@ -285,6 +291,9 @@ class Session(requests.Session):
         port=None,
         verify_ssl=None,
         token=None,
+        client_id=None,
+        client_secret=None,
+        consul_token=None,
     ):
         super(Session, self).__init__()
 
@@ -295,9 +304,6 @@ class Session(requests.Session):
 
         self._id = uuid4().hex
         self.message_log = logger.getChild("session.%s" % self._id)
-
-        if verify_ssl is None:
-            verify_ssl = True
 
         # If certificate path has already been set for SWAT package, make
         # Requests module reuse it.
@@ -376,40 +382,56 @@ class Session(requests.Session):
             "password": password,
         }
 
-        if self._settings["password"] is None:
+        if password is None and client_secret is None:
             # Try to get credentials from .authinfo or .netrc files.
             # If no file path was specified, the default locations will
             # be checked.
-            if "swat" in sys.modules:
+            try:
                 auth = swat.utils.authinfo.query_authinfo(
                     domain, user=username, path=authinfo
                 )
-                self._settings["username"] = auth.get("user")
-                self._settings["password"] = auth.get("password")
+                if auth:
+                    self._settings["username"] = auth.get("user")
+                    self._settings["password"] = auth.get("password")
+            except AttributeError:
+                # If swat package or authinfo module not available
+                pass
 
             # Not able to load credentials using SWAT.  Try Netrc.
-            # TODO: IF a username was specified, verify that the credentials
-            #       found are for that username.
             if self._settings["password"] is None:
                 try:
                     parser = netrc.netrc(authinfo)
                     values = parser.authenticators(domain)
                     if values:
-                        (
-                            self._settings["username"],
-                            _,
-                            self._settings["password"],
-                        ) = values
+                        found_user, _, found_pass = values
+
+                        # Only use the credentials if they match the specified user (if
+                        # one was specified).
+                        if username is None or username.lower() == found_user:
+                            self._settings["username"] = found_user
+                            self._settings["password"] = found_pass
                 except (OSError, IOError):
                     pass  # netrc throws if $HOME is not set
 
         # Set this prior authentication attempts
         self.verify = verify_ssl
 
-        # Find a suitable authentication mechanism and build an auth header
-        self.auth = self.get_auth(
-            self._settings["username"], self._settings["password"], token
-        )
+        if consul_token:
+            self.auth = self._request_token_with_consul(
+                consul_token, client_id=client_id
+            )
+        else:
+            # Find a suitable authentication mechanism and build an auth header
+            self.auth = self._get_authorization_token(
+                token=token,
+                username=self.username,
+                password=self._settings["password"],
+                client_id=client_id,
+                client_secret=client_secret,
+            )
+
+        # Used to cache version info
+        self._version_info = None
 
         # Used for context manager
         self._old_session = current_session()
@@ -511,11 +533,11 @@ class Session(requests.Session):
 
         # Starting in SWAT v1.8 oauth tokens could be passed directly in the password param.
         # Otherwise, use the username & password to re-authenticate.
+        # Use this sessions info to connect to CAS unless user has explicitly give a value (even if None)
         if version.parse(swat.__version__) >= version.parse("1.8"):
-            kwargs["username"] = None
-            kwargs["password"] = self.auth.access_token
+            kwargs.setdefault("username", None)
+            kwargs.setdefault("password", self.auth.access_token)
         else:
-            # Use this sessions info to connect to CAS unless user has explicitly give a value (even if None)
             kwargs.setdefault("username", self.username)
             kwargs.setdefault("password", self._settings["password"])
 
@@ -634,7 +656,7 @@ class Session(requests.Session):
                 # Access token expired, need to refresh it (if we can)
                 if "access token expired" in auth_header:
                     try:
-                        self.auth = self.get_oauth_token(
+                        self.auth = self._request_token_with_oauth(
                             refresh_token=self.auth.refresh_token
                         )
 
@@ -685,191 +707,6 @@ class Session(requests.Session):
 
     def delete(self, url, **kwargs):
         return self.request("DELETE", url, **kwargs)
-
-    def get_auth(self, username=None, password=None, token=None):
-        """Attempt to authenticate with the Viya environment.
-
-        If `username` and `password` or `token` are provided, they will be used exclusively.  If neither of these
-        is specified, Kerberos authorization will be attempted.  If this fails, authorization using an
-        OAuth2 authorization code will be attempted.  Be aware that this requires prompting the user for input and
-        should NOT be used in a non-interactive session.
-
-        Parameters
-        ----------
-        username : str, optional
-            Username to use for authentication.
-        password : str, optional
-            Password to use for authentication.  Required if `username` is provided.
-        token : str, optional
-            An existing OAuth2 access token to use.
-
-        Returns
-        -------
-        OAuth2Token
-            Authorization header to use with requests to the evironent
-
-        """
-
-        # If explicit username & password were provided, use them.
-        if username is not None and password is not None:
-            return self.get_oauth_token(username, password)
-
-        # If an existing access token was provided, use it
-        if token is not None:
-            return OAuth2Token(token)
-
-        # Kerberos doesn't require any interruption to the user, so try that first if username/password
-        # were not provided.
-        try:
-            token = self._get_token_with_kerberos()
-            return OAuth2Token(token)
-        except:
-            logger.exception("Failed to connect with Kerberos")
-
-        # Try loading cached credentials
-        token = self.read_cached_token(self.PROFILE_PATH)
-
-        if token is not None:
-            return token
-
-        # If we got this far, then no password and no kerberos.  Try prompting the user for an authorization code.
-        auth_code = self.prompt_for_auth_code()
-        return self.get_oauth_token(auth_code=auth_code)
-
-    def get_oauth_token(
-        self,
-        username=None,
-        password=None,
-        auth_code=None,
-        refresh_token=None,
-        client_id=None,
-        client_secret=None,
-    ):
-        """Request an OAuth2 access token using either a username & password or an auth token.
-
-        Parameters
-        ----------
-        username : str, optional
-            Username to use for authentication.
-        password : str, optional
-            Password to use for authentication.  Required if `username` is provided.
-        auth_code : str, optional
-            Authorization code to use.
-        refresh_token : str, optinoal
-            Valid refresh token to use.
-        client_id : str, optional
-            Client ID requesting access.  Use if connection to Viya should be made using a non-default client id.
-        client_secret : str, optional
-            Client secret for client requesting access.  Required if `client_id` is provided
-
-        Returns
-        -------
-        OAuth2Token
-
-        See Also
-        --------
-        :meth:`Session.prompt_for_auth_code`
-
-        """
-
-        if username is not None:
-            data = {
-                "grant_type": "password",
-                "username": username,
-                "password": password,
-            }
-        elif auth_code is not None:
-            data = {"grant_type": "authorization_code", "code": auth_code}
-        elif refresh_token is not None:
-            data = {"grant_type": "refresh_token", "refresh_token": refresh_token}
-        else:
-            raise ValueError(
-                "Either username and password or auth_code parameters must be specified."
-            )
-
-        client_id = client_id or os.environ.get("SASCTL_CLIENT_ID", "sas.ec")
-        client_secret = client_secret or os.environ.get("SASCTL_CLIENT_SECRET", "")
-
-        headers = {
-            "Accept": "application/json",
-            "Content-Type": "application/x-www-form-urlencoded",
-        }
-
-        url = self._build_url("/SASLogon/oauth/token")
-        r = super(Session, self).post(
-            url,
-            headers=headers,
-            data=data,
-            auth=(client_id, client_secret),
-            verify=self.verify,
-        )
-
-        # Raise user-friendly error messages if issue is known
-        if r.status_code == 400 and auth_code is not None:
-            j = r.json()
-            if "Invalid authorization code" in j.get("error_description", ""):
-                raise exceptions.AuthorizationError(
-                    "Invalid authorization code: '%s'" % auth_code
-                )
-        if r.status_code == 401:
-            if (
-                r.json().get("error_description", "").lower() == "bad credentials"
-                and username is None
-            ):
-                raise exceptions.AuthenticationError(msg="Invalid client id or secret.")
-            if r.json().get("error", "") == "invalid_token":
-                raise exceptions.AuthorizationError(
-                    "Refresh token is incorrect, expired, or revoked."
-                )
-            if username is not None:
-                raise exceptions.AuthenticationError(username)
-
-        r.raise_for_status()
-
-        token = OAuth2Token(**r.json())
-
-        # No reason to cache token if username & password were provided - they'd just be re-provided on future
-        # connections.  Cache for auth code & refresh token to prevent frequent interruptions for the user.
-        if username is None:
-            self.cache_token(token, self.PROFILE_PATH)
-
-        return token
-
-    def prompt_for_auth_code(self, client_id=None):
-        """Prompt the user open a URL to generate an auth code.
-
-        Note that this halts program execution until input is received and should only be used for interactive sessions.
-
-        Parameters
-        ----------
-        client_id : str, optional
-
-        Returns
-        -------
-        str
-            Authorization code that can be used to acquire an OAuth2 access token.
-
-        See Also
-        --------
-        :meth:`Session.get_oauth_token`
-
-        """
-        client_id = client_id or os.environ.get("SASCTL_CLIENT_ID", "sas.ec")
-
-        # User must open this URL in a browser and then enter the auth code that's generated.
-        url = (
-            self._build_url("/SASLogon/oauth/authorize")
-            + "?response_type=code&client_id="
-            + client_id
-        )
-        message = (
-            "Please use a web browser to login at the following URL to get your authorization code:\n"
-            + url
-        )
-        print(message)
-        auth_code = input("Authorization Code:")
-
-        return auth_code
 
     def cache_token(self, token, path):
         """Write an OAuth2 token to the cache.
@@ -960,7 +797,9 @@ class Session(requests.Session):
                 # Attempt to refresh if cached token has expired
                 if token.is_expired:
                     try:
-                        token = self.get_oauth_token(refresh_token=token.refresh_token)
+                        token = self._request_token_with_oauth(
+                            refresh_token=token.refresh_token
+                        )
                     except (
                         exceptions.AuthorizationError,
                         requests.exceptions.HTTPError,
@@ -971,23 +810,362 @@ class Session(requests.Session):
                 if not token.is_expired:
                     return token
 
-    def _get_token_with_kerberos(self):
-        """Authenticate with a Kerberos ticket."""
+    def version_info(self):
+        """Get version information from the connected SAS Viya environment
+
+        Returns
+        -------
+        VersionInfo
+
+        Notes
+        -----
+        The resulting version information is cached and returned on any subsequent calls.  This
+        allows repeatedly checking version information without making redundant network calls to the
+        SAS Viya server.
+
+        """
+        # The Viya environment isn't changing, so there's no need to repeatedly make
+        # service calls to check version information.  If we've already cached the info
+        # then just return it.
+        if self._version_info:
+            return self._version_info
+
+        try:
+            # Try to determine if we're talking to Viya 3 or 4
+            r = self.get("/licenses/grants")
+            release = r.json().get("release")
+
+            # Convert 'V03' and 'V04' to just 3 or 4.
+            major_version = int(release.upper().lstrip("V"))
+
+            # No good way to get detailed version info from a Viya 3 environment.
+            # At this point, we just assume it's Viya 3.5 and return
+            if major_version == 3:
+                self._version_info = VersionInfo(major_version)
+            else:
+                # Endpoint with detailed release info only available for Viya 4
+                cadence_info = self.get("/deploymentData/cadenceVersion").json()
+                name = cadence_info["cadenceName"]
+                release = cadence_info["cadenceVersion"]
+                self._version_info = VersionInfo(
+                    major_version, cadence=name, release=release
+                )
+        except HTTPError:
+            # Ignore.  We'll return None and (possibly) replace with correct info on subsequent call.
+            pass
+
+        return self._version_info
+
+    def _build_url(self, url):
+        """Build a complete URL from a path by substituting in session parameters."""
+        components = urlsplit(url)
+
+        domain = components.netloc or self._settings["domain"]
+
+        # Add port if a non-standard port was specified
+        if self._settings["port"] is not None and ":" not in domain:
+            domain = f"{domain}:{self._settings['port']}"
+
+        return urlunsplit(
+            [
+                components.scheme or self._settings["protocol"],
+                domain,
+                components.path,
+                components.query,
+                components.fragment,
+            ]
+        )
+
+    def _get_authorization_token(
+        self,
+        token=None,
+        username=None,
+        password=None,
+        client_id=None,
+        client_secret=None,
+    ):
+        """Authenticate with SAS Viya and obtain and access token.
+
+        This method supports multiple authentication methods:
+
+         - an existing OAuth2 token
+         - password authentication
+         - client credentials
+         - Kerberos
+         - cached tokens (from previous authorization codes)
+         - authorization code
+
+        If authentication using client credentials fails because the
+        client_credentials grant type is not allowed the token cache will
+        be searched for a valid token.  If no valid token is found, the
+        user will be prompted to generate and enter an authorization code.
+
+        Parameters
+        ----------
+        token : str, optional
+            An existing access token to reuse.
+        username : str, optional
+            Name of the user account to use.  Used for password
+            authentication and Kerberos authentication.
+        password : str, optional
+            Password corresponding to `username`.  Only used for password
+            authentication.
+        client_id : str, optional
+            The id of the client to use during authentication.  Overrides
+            the SASCTL_CLIENT_ID environment variable.  Used for password,
+            client_credentials, and authorization_code access.
+        client_secret : str, optional
+            The client secret to use during authentication.  Overrides the
+            SASCTL_CLIENT_SECRET environment variable.  Used for password,
+            client_credentials, and authorization_code access.
+
+        Returns
+        -------
+        OAuth2Token
+
+        Raises
+        ------
+        AuthenticationError
+            For invalid username/password combination or invalid client id
+            and client secret combination.
+
+        AuthorizationError
+            If authorization code is invalid or if refresh token is expired.
+
+        ValueError
+            If a client id is specified (either through `client_id` or
+            SASCTL_CLIENT ID) but no client secret is provided.
+
+        """
+        if token:
+            return OAuth2Token(token)
+
+        if username and password:
+            return self._request_token_with_oauth(
+                username, password, client_id, client_secret
+            )
+
+        client_id_provided = (
+            client_id is not None or os.getenv("SASCTL_CLIENT_ID") is not None
+        )
+        client_secret_provided = (
+            client_secret is not None or os.getenv("SASCTL_CLIENT_SECRET") is not None
+        )
+
+        # Attempt authentication with just client credentials
+        if client_id_provided:
+            if not client_secret_provided:
+                # Authenticating with client credentials should require a client secret.
+                raise ValueError(
+                    "A client secret must be provided whenever a client id is specified."
+                )
+
+            try:
+                return self._request_token_with_oauth(
+                    client_id=client_id, client_secret=client_secret
+                )
+            except RuntimeError:
+                # If all we have is a client id & client password, there's no way to
+                # know whether we're supposed to authenticate directly using the client
+                # credentials, or prompt the user for an authorization code.  Try
+                # client credential authentication first since it doesn't require user
+                # prompts, but if that fails do not raise an exception, we'll try
+                # requesting an authorization code.
+                logger.debug("Authentication using client credentials is disallowed.")
+
+        # If the necessary Python package is installed try to get a token using Kerberos.
+        # NOTE: username may be None since client may already have valid, cached
+        #       Kerberos tickets.  If so, Session.username will be updated with the
+        #       username from Kerberos.
+        if kerberos:
+            try:
+                return self._request_token_with_kerberos(username)
+            except:
+                logger.exception(
+                    "Encountered an error while attempting Kerberos authorization."
+                )
+        else:
+            logger.debug(
+                "Skipping Kerberos authentication - kerberos and winkerberos "
+                "packages not found."
+            )
+
+        # Before we prompt the user for an authorization code, check if there's already
+        # a valid token in the cache from a previous session.
+        token = self.read_cached_token(self.PROFILE_PATH)
+
+        if token is not None:
+            return token
+
+        # If we got this far, then no password and no kerberos.  Try prompting the user
+        # for an authorization code.
+        auth_code = self._prompt_for_auth_code(client_id)
+        token = self._request_token_with_oauth(
+            client_id=client_id, client_secret=client_secret, auth_code=auth_code
+        )
+
+        # Cache the token so we don't have to request user input again until the
+        # token expires.
+        self.cache_token(token, self.PROFILE_PATH)
+
+        return token
+
+    def _prompt_for_auth_code(self, client_id=None):
+        """Prompt the user open a URL to generate an auth code.
+
+        Note that this halts program execution until input is received and
+        should only be used for interactive sessions.
+
+        Parameters
+        ----------
+        client_id : str, optional
+
+        Returns
+        -------
+        str
+            Authorization code that can be used to acquire an OAuth2
+            access token.
+
+        See Also
+        --------
+        :meth:`Session.get_oauth_token`
+
+        """
+        client_id = client_id or os.environ.get("SASCTL_CLIENT_ID", "sas.ec")
+
+        # User must open this URL in a browser and enter the auth code that's generated.
+        url = (
+            self._build_url("/SASLogon/oauth/authorize")
+            + "?response_type=code&client_id="
+            + client_id
+        )
+        message = (
+            f"Please use a web browser to login at the following URL to get your "
+            f"authorization code:\n{url}"
+        )
+        print(message)
+        auth_code = input("Authorization Code:")
+
+        return auth_code
+
+    @staticmethod
+    def _read_token_cache(path):
+        """Read cached OAuth2 access tokens from disk.
+
+        Parameters
+        ----------
+        path : str
+            Path to file containing cached tokens.
+
+        Returns
+        -------
+        dict or None
+
+        Raises
+        ------
+        RuntimeError
+            If file permissions are too permissive.
+
+        """
+        yaml_file = os.path.expanduser(path)
+
+        # See if a token has been cached for the hostname
+        if os.path.exists(yaml_file):
+            # Get bit flags indicating access permissions
+            mode = os.stat(yaml_file).st_mode
+            flags = oct(mode)[-3:]
+
+            if flags != "600":
+                raise RuntimeError(
+                    f"Unable to read profile cache.  The file permissions for "
+                    f"{yaml_file} must be configured so that only the file owner has "
+                    f"read/write permissions (equivalent to 600 on Linux systems)."
+                )
+
+            with open(yaml_file) as f:
+                return yaml.safe_load(f)
+
+        return None
+
+    def _request_token_with_consul(self, consul_token, client_id=None):
+        """Request an OAuth token from Consul.
+
+        This functionality is reserved for system administrators as access
+        to the Consul token is restricted.
+
+        Parameters
+        ----------
+        consul_token : str
+            A Consul ACL token
+        client_id : str, optional
+            The name of the client being used to access the service.
+
+        Returns
+        -------
+        OAuth2Token
+
+        Raises
+        ------
+        AuthenticationError
+            If unable to retrieve an access token for any reason.
+
+        """
+        client_id = client_id or ""
+
+        headers = {"X-Consul-Token": consul_token}
+        params = {"callback": False, "serviceId": client_id}
+
+        url = self._build_url("/SASLogon/oauth/clients/consul")
+        response = super(Session, self).post(
+            url,
+            headers=headers,
+            params=params,
+            verify=self.verify,
+        )
+
+        try:
+            response.raise_for_status()
+            data = response.json()
+            return OAuth2Token(**data)
+        except requests.HTTPError as e:
+            raise exceptions.AuthenticationError(
+                msg="Failed to authenticate using Consul token."
+            ) from e
+
+    def _request_token_with_kerberos(self, username=None):
+        """Authenticate with a Kerberos ticket.
+
+        Parameters
+        ----------
+        username : str, optional
+
+        Returns
+        -------
+        OAuth2Token
+
+        Raises
+        ------
+        RuntimeError
+            If required kerberos package is not installed.
+
+        ValueError
+            If there is any issue with the server's response to the
+            authorization request.
+
+        """
         if kerberos is None:
             raise RuntimeError(
                 "Kerberos package not found.  Run 'pip "
                 "install sasctl[kerberos]' to install."
             )
 
-        user = self.username
-        client_id = "sas.tkmtrb"
         flags = kerberos.GSS_C_MUTUAL_FLAG | kerberos.GSS_C_SEQUENCE_FLAG
         service = "HTTP@%s" % self._settings["domain"]
 
         logger.info("Attempting Kerberos authentication to %s", service)
 
         url = self._build_url(
-            "/SASLogon/oauth/authorize?client_id=%s&response_type=token" % client_id
+            "/SASLogon/oauth/authorize?client_id=sas.tkmtrb&response_type=token"
         )
 
         # Get Kerberos challenge
@@ -1013,7 +1191,9 @@ class Session(requests.Session):
             )
 
         # Initialize a request to KDC for a ticket to access the service.
-        _, context = kerberos.authGSSClientInit(service, principal=user, gssflags=flags)
+        _, context = kerberos.authGSSClientInit(
+            service, principal=username, gssflags=flags
+        )
 
         # Send the request.
         # NOTE: empty-string parameter required for initial call.
@@ -1050,65 +1230,169 @@ class Session(requests.Session):
                 "header does not contain an access token."
             )
 
-        return match.group(0)
+        # Extract access token and return as an Oauth token
+        return OAuth2Token(match.group(0))
 
-    def _build_url(self, url):
-        """Build a complete URL from a path by substituting in session parameters."""
-        components = urlsplit(url)
+    def _request_token_with_oauth(
+        self,
+        username=None,
+        password=None,
+        client_id=None,
+        client_secret=None,
+        auth_code=None,
+        refresh_token=None,
+    ):
+        """Request a token from the SAS SASLogon service.
 
-        domain = components.netloc or self._settings["domain"]
-
-        # Add port if a non-standard port was specified
-        if self._settings["port"] is not None and ":" not in domain:
-            domain = "{}:{}".format(domain, self._settings["port"])
-
-        return urlunsplit(
-            [
-                components.scheme or self._settings["protocol"],
-                domain,
-                components.path,
-                components.query,
-                components.fragment,
-            ]
-        )
-
-    @staticmethod
-    def _read_token_cache(path):
-        """Read cached OAuth2 access tokens from disk.
+        Supports four different flows:
+         - authenticate with a username & password and receive a token
+         - authenticate with a client id & secret and receive a token
+         - provide an authorization code and receive a token
+         - provide a refresh token and receive a new token
 
         Parameters
         ----------
-        path : str
-            Path to file containing cached tokens.
+        username : str, optional
+            Username of the user in SAS Viya.  Required for password
+            authentication flow.
+        password : str, optional
+            Password of the user in SAS Viya.  Required for password
+            authentication flow.
+        client_id : str, optional
+            Client ID of a client registered with SAS Viya.  Required for
+            client credentials flow, but will use
+            a default client ID if not provided during password or
+            authorization code flows.
+        client_secret : str, optional
+            Client secret of a client registered with SAS Viya.  Required
+            for client credentials flow, but will use a default if not
+            provided during password or authorization code flows.
+        auth_code : str, optional
+            An authorization code obtained by the user after
+            authenticating with SAS Viya.  Required for authorization
+            code flow.
+        refresh_token : str, optional
+            A refresh token obtained during a previous authorization
+            request.  Required if requesting a refreshed access token.
 
         Returns
         -------
-        dict or None
+        OAuth2Token
 
         Raises
         ------
+        ValueError
+            If no valid parameter combination is provided.
+
         RuntimeError
-            If file permissions are too permissive.
+            If the requested grant type is not allowed for the given
+            `client_id`.
+
+        AuthenticationError
+            For invalid username/password combination or invalid
+            client id/secret combination.
+
+        AuthorizationError
+            If authorization code is invalid or if refresh token is expired.
+
+        requests.HTTPError
+            For any unexpected/unhandled HTTP error code (e.g. HTTP 404)
+
+        requests.ConnectionError
+            If there's an issue establishing a connection with the server.
 
         """
-        yaml_file = os.path.expanduser(path)
+        client_id = client_id or os.getenv("SASCTL_CLIENT_ID", "sas.ec")
+        client_secret = client_secret or os.getenv("SASCTL_CLIENT_SECRET", "")
 
-        # See if a token has been cached for the hostname
-        if os.path.exists(yaml_file):
-            # Get bit flags indicating access permissions
-            mode = os.stat(yaml_file).st_mode
-            flags = oct(mode)[-3:]
+        # Order in which parameters are checked is important.  Passing client
+        # credentials could indicate password, auth code, or client credential
+        # authentication.  Only use client credential flow if first two are
+        # ruled out.
+        if username:
+            anchor = "#password"
+            data = {
+                "grant_type": "password",
+                "username": username,
+                "password": password,
+            }
+            logger.debug("Attempting password authentication as user '%s'.", username)
+        elif auth_code:
+            anchor = "#authorization_code"
+            data = {"grant_type": "authorization_code", "code": auth_code}
+            logging.debug("Attempting authorization using an auth code.")
+        elif client_secret:
+            anchor = "#client_credentials"
+            data = {"grant_type": "client_credentials"}
+            logger.debug(
+                "Attempting client credential authentication as client '%s'.", client_id
+            )
+        elif refresh_token:
+            anchor = ""
+            data = {"grant_type": "refresh_token", "refresh_token": refresh_token}
+        else:
+            raise ValueError("At least one set of parameters must be provided.")
 
-            if flags != "600":
-                raise RuntimeError(
-                    "Unable to read profile cache.  "
-                    "The file permissions for %s must be configured so that only the file owner has "
-                    "read/write permissions (equivalent to 600 on Linux systems)."
-                    % yaml_file
+        headers = {
+            "Accept": "application/json",
+            "Content-Type": "application/x-www-form-urlencoded",
+        }
+
+        url = self._build_url(f"/SASLogon/oauth/token{anchor}")
+        response = super(Session, self).post(
+            url,
+            headers=headers,
+            data=data,
+            auth=(client_id, client_secret),
+            verify=self.verify,
+        )
+
+        try:
+            data = response.json()
+        except json.JSONDecodeError:
+            # A valid response (and some HTTP errors) should contain valid JSON.  In
+            # the extremely unlikely event that an HTTP error is returned that doesn't
+            # include JSON (e.g. HTTP 404) just raise the HTTP error.
+            response.raise_for_status()
+
+        # If request failed for a known reason, raise a user-friendly error message.
+        if (
+            response.status_code == 400
+            and auth_code is not None
+            and "Invalid authorization code" in data.get("error_description", "")
+        ):
+            raise exceptions.AuthorizationError(
+                f"Invalid authorization code: {auth_code}."
+            )
+
+        if response.status_code == 401:
+            # Response is the same if either username/password or client id/secret is
+            # invalid ('{"error":"unauthorized","error_description":"Bad credentials"}')
+            # We're assuming that if a username was provided, that's probably the problem.
+            # NOTE: this does mean that valid credentials with incorrectly set client
+            #       id & secret results in an error message that blames the username
+            #       and password combination.
+            if "bad credentials" in data.get("error_description", "").lower():
+                if username is not None:
+                    raise exceptions.AuthenticationError(username)
+
+                raise exceptions.AuthenticationError(msg="Invalid client id or secret.")
+
+            # If client used is not allowed to authenticate using the requested method
+            # an error is returned.  Example:
+            # '{"error":"invalid_client","error_description":"Unauthorized grant type: password"}'
+            if "unauthorized grant type" in data.get("error_description", "").lower():
+                raise RuntimeError(data["error_description"])
+
+            if refresh_token is not None and data.get("error", "") == "invalid_token":
+                raise exceptions.AuthorizationError(
+                    "Refresh token is incorrect, expired, or revoked."
                 )
 
-            with open(yaml_file) as f:
-                return yaml.load(f, Loader=yaml.FullLoader)
+        # Raise a generic exception for any other issues
+        response.raise_for_status()
+
+        return OAuth2Token(**data)
 
     @staticmethod
     def _write_token_cache(profiles, path):
@@ -1211,7 +1495,7 @@ class PageIterator:
             self._min_queue_len = 0
             self._start = 0
             self._limit = 0
-        if link is not None:
+        else:
             link = link["href"]
             start = re.search(r"(?<=start=)[\d]+", link)
             limit = re.search(r"(?<=limit=)[\d]+", link)
@@ -1326,7 +1610,8 @@ class PagedItemIterator:
         # Total number of items to iterate over
         if "count" in obj:
             # NOTE: "count" may be an (over) estimate of the number of items available
-            #       since some may be inaccessible due to user permissions & won't actually be returned.
+            #       since some may be inaccessible due to user permissions & won't
+            #       actually be returned.
             self._count = int(obj.count)
         else:
             self._count = len(obj["items"])
@@ -1468,6 +1753,125 @@ class PagedList(list):
             string = string.rstrip("]") + ", ...]"
 
         return string
+
+
+class VersionInfo:
+    """Stores the version information for a SAS Viya environment.
+
+    Parameters
+    ----------
+    major : int
+        Major version number (e.g. 3 for Viya 3.5 or 4 for Viya 4.0)
+    minor : int
+        Minor version number (e.g. 5 for Viya 3.5 or 0 for Viya 4.0)
+    cadence : str, optional
+        Release cadence for Viya 4.  Should be one of 'stable' or 'LTS'.
+    release : str, optional
+        Release number for Viya 4.  Two formats are currently possible:
+         - YYYY.R.U where R is the LTS release number in YYYY and U is the updates since R
+         - YYYY.MM where MM is the month of the release.
+
+    """
+
+    def __init__(self, major, minor=None, cadence=None, release=None):
+        self._major = major
+        self._minor = minor if minor else 5 if major == 3 else 0
+        self._cadence = str(cadence) if cadence else None
+        self._release = str(release) if release else None
+
+    def __ge__(self, other):
+        return self > other or self == other
+
+    def __gt__(self, other):
+        return self._compare(other) > 0
+
+    def __lt__(self, other):
+        return self._compare(other) < 0
+
+    def __le__(self, other):
+        return self < other or self == other
+
+    def __eq__(self, other):
+        return self._compare(other) == 0
+
+    def __float__(self):
+        return float(self._major) + 0.1 * (self._minor or 0)
+
+    def __hash__(self):
+        return hash((self.major, self.minor, self.release))
+
+    def _compare(self, other):
+        """Compare and return -1/0/1 indicating lt/eq/gt.
+
+        Parameters
+        ----------
+        other : any
+
+        Returns
+        -------
+        int
+            -1 if < `other`, 0 if equal, and 1 if > `other`.
+
+        """
+        # Compare Major/Minor versions (e.g. Viya 3.5 < Viya 4.0)
+        if float(self) > float(other):
+            return 1
+        if float(self) < float(other):
+            return -1
+
+        # If comparing two Viya 4 versions, may need to check actual release number  to determine order
+        if self.release and getattr(other, "release", None):
+            if self._release == other.release:
+                return 0
+
+            parts = self.release.split(".")
+            other_parts = other.release.split(".")
+
+            # Release format was changed from YYYY.r.u to YYYY.MM so any release with 2 '.' is older than
+            # a release with 1 '.'
+            if len(parts) == 2 and len(other_parts) == 3:
+                return 1
+            if len(parts) == 3 and len(other_parts) == 2:
+                return -1
+
+            # If we got this far then both version numbers should have the same release format.
+            # Earlier release is the one with at least one component part that is lower
+            for a, b in zip(parts, other_parts):
+                if int(a) < int(b):
+                    return -1
+                if int(a) > int(b):
+                    return 1
+
+        # Either other doesn't have a .release or we compared all components of .release and
+        # they were all equal
+        return 0
+
+    def __repr__(self):
+        name = f"{self.__class__.__name__}(major={self._major}, minor={self._minor}"
+
+        if self._cadence:
+            name += ", cadence='%s'" % self._cadence
+
+        if self._release:
+            name += ", release='%s'" % self._release
+
+        return name + ")"
+
+    @property
+    def cadence(self):
+        return self._cadence
+
+    @property
+    def major(self):
+        return self._major
+
+    @property
+    def minor(self):
+        return self._minor
+
+    @property
+    def release(self):
+        return self._release
 
 
 def is_uuid(id_):
@@ -1949,6 +2353,12 @@ def platform_version():
         SAS Viya version number
 
     """
+    warnings.warn(
+        "platform_version() has been deprecated and will be removed in a future version.  "
+        "Please use Session.version_info() instead.",
+        DeprecationWarning,
+    )
+
     from .services import model_repository as mr
 
     response = mr.info()
