@@ -4,4 +4,9 @@
 # Copyright © 2019, SAS Institute Inc., Cary, NC, USA.  All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-from .astore import create_package, create_package_from_astore
+from .astore import (
+    create_package,
+    create_package_from_astore,
+    create_package_from_datastep,
+)
+from .modelMigration import convertModelZip
