@@ -12,37 +12,34 @@ from sasctl.core import request_link
 from sasctl.services import concepts as cp
 
 
-pytestmark = pytest.mark.usefixtures('session')
+pytestmark = pytest.mark.usefixtures("session")
 
 
 def assert_job_succeeds(job):
-    assert job.state == 'pending'
+    assert job.state == "pending"
 
-    while request_link(job, 'state') in ('pending', 'running'):
+    while request_link(job, "state") in ("pending", "running"):
         time.sleep(1)
 
-    state = request_link(job, 'state')
+    state = request_link(job, "state")
 
-    if state == 'failed':
+    if state == "failed":
         # Refresh to get 'errors' ref
-        job = request_link(job, 'self')
-        errors = request_link(job, 'errors')
-        pytest.fail('Job failed: ' + str(errors))
-    assert state == 'completed'
+        job = request_link(job, "self")
+        errors = request_link(job, "errors")
+        pytest.fail("Job failed: " + str(errors))
+    assert state == "completed"
 
 
 def test_from_table(cas_session, airline_dataset):
-    TABLE_NAME = 'airline_tweets'
-    cas_session.upload(airline_dataset, casout=dict(name=TABLE_NAME,
-                                                    replace=True))
+    TABLE_NAME = "airline_tweets"
+    cas_session.upload(airline_dataset, casout=dict(name=TABLE_NAME, replace=True))
     from sasctl.services import cas_management as cm
 
-    cas_session.table.promote(TABLE_NAME, targetlib='Public')
+    cas_session.table.promote(TABLE_NAME, targetlib="Public")
 
-    input = cm.get_table(TABLE_NAME, 'Public')
-    job = cp.assign_concepts(input,
-                               id_column='tweet_id',
-                               text_column='text')
+    input = cm.get_table(TABLE_NAME, "Public")
+    job = cp.assign_concepts(input, id_column="tweet_id", text_column="text")
 
     assert_job_succeeds(job)
 
@@ -50,7 +47,7 @@ def test_from_table(cas_session, airline_dataset):
 def test_from_inline_docs():
     from sasctl.services import cas_management as cm
 
-    caslib = cm.get_caslib('Public')
+    caslib = cm.get_caslib("Public")
     input = [
         "Oh yes, the, uh, the Norwegian Blue. What's wrong with it?",
         "I'll tell you what's wrong with it, my lad. He's dead, that's "
@@ -60,7 +57,7 @@ def test_from_inline_docs():
         "at one right now.",
         "No no he's not dead, he's, he's resting! Remarkable bird, "
         "the Norwegian Blue, isn't it? Beautiful plumage!",
-        "The plumage don't enter into it. It's stone dead."
+        "The plumage don't enter into it. It's stone dead.",
     ]
 
     job = cp.assign_concepts(input, caslib=caslib)
