@@ -6,10 +6,11 @@
 
 import pytest
 
+from sasctl import current_session
 from sasctl.services import model_publish as mp
 
 # Every test function in the module will automatically receive the session fixture
-pytestmark = pytest.mark.usefixtures('session')
+pytestmark = pytest.mark.usefixtures("session")
 
 
 @pytest.mark.incremental
@@ -18,37 +19,46 @@ class TestModelPublish:
         destinations = mp.list_destinations()
 
         assert isinstance(destinations, list)
-        assert any(d.name == 'maslocal' for d in destinations)
+        assert any(d.name == "maslocal" for d in destinations)
 
     def test_get_publish_destination(self):
-        dest = mp.get_destination('maslocal')
+        dest = mp.get_destination("maslocal")
 
-        assert dest.name == 'maslocal'
-        assert dest.destinationType == 'microAnalyticService'
+        assert dest.name == "maslocal"
+        assert dest.destinationType == "microAnalyticService"
 
     def test_create_cas_destination(self):
-        dest = mp.create_cas_destination('sasctlcas', 'Public', 'sasctl_models',
-                                         description='Test CAS publish destination from sasctl.')
+        dest = mp.create_cas_destination(
+            "sasctlcas",
+            "Public",
+            "sasctl_models",
+            description="Test CAS publish destination from sasctl.",
+        )
 
-        assert dest.name == 'sasctlcas'
-        assert dest.destinationType == 'cas'
-        assert dest.casLibrary == 'Public'
-        assert dest.casServerName == 'cas-shared-default'
-        assert dest.destinationTable == 'sasctl_models'
-        assert dest.description == 'Test CAS publish destination from sasctl.'
+        assert dest.name == "sasctlcas"
+        assert dest.destinationType == "cas"
+        assert dest.casLibrary == "Public"
+        assert dest.casServerName == "cas-shared-default"
+        assert dest.destinationTable == "sasctl_models"
+        assert dest.description == "Test CAS publish destination from sasctl."
+
+    def test_delete_cas_destination(self):
+        dest = mp.get_destination("sasctlcas")
+        assert dest.name == "sasctlcas"
+
+        mp.delete_destination("sasctlcas")
+
+        dest = mp.get_destination("sasctlcas")
+        assert dest is None
 
     def test_create_mas_destination(self):
-        dest = mp.create_mas_destination('sasctlmas', 'localhost')
+        if current_session().version_info() == 4:
+            pytest.skip(
+                "Publishing destinations for a remote SAS Micro Analytic Service are currently not supported."
+            )
 
-        assert dest.name == 'sasctlmas'
-        assert dest.destinationType == 'microAnalyticService'
-        assert 'description' not in dest
+        dest = mp.create_mas_destination("sasctlmas", "localhost")
 
-    def test_delete_cas_destionation(self):
-        dest = mp.get_destination('sasctlcas')
-        assert dest.name == 'sasctlcas'
-
-        mp.delete_destination('sasctlcas')
-
-        dest = mp.get_destination('sasctlcas')
-        assert dest is None
+        assert dest.name == "sasctlmas"
+        assert dest.destinationType == "microAnalyticService"
+        assert "description" not in dest
