@@ -6,13 +6,14 @@
 
 import pytest
 
-from sasctl import RestObj
+from sasctl import current_session, RestObj
 from sasctl.services import projects as proj
 
 
-pytestmark = pytest.mark.usefixtures('session')
+pytestmark = pytest.mark.usefixtures("session")
 
-PROJECT_NAME = 'Test Project'
+PROJECT_NAME = "Test Project"
+
 
 def test_list_projects():
     projects = proj.list_projects()
@@ -20,11 +21,14 @@ def test_list_projects():
 
 
 def test_get_project():
-    project = proj.get_project('Not a Project')
+    project = proj.get_project("Not a Project")
     assert project is None
 
 
 def test_create_project():
+    if current_session().version_info() >= 4:
+        pytest.skip("Projects service was removed from Viya 4.")
+
     project = proj.create_project(PROJECT_NAME)
     assert isinstance(project, RestObj)
     assert PROJECT_NAME == project.name
