@@ -9,12 +9,12 @@ import pytest
 from sasctl import RestObj
 from sasctl.services import reports, report_images
 
-pytestmark = pytest.mark.usefixtures('session')
+pytestmark = pytest.mark.usefixtures("session")
 
 
 @pytest.mark.incremental
 class TestReportImages:
-    REPORT_NAME = 'CAS Activity'
+    REPORT_NAME = "Retail Insights"
 
     def test_get_report(self, request):
         """Verify that the report exists before using it in tests."""
@@ -23,10 +23,10 @@ class TestReportImages:
         assert report.name == self.REPORT_NAME
 
         # Store the report for subsequent tests
-        request.config.cache.set('REPORT', report)
+        request.config.cache.set("REPORT", report)
 
     def test_get_images(self, request):
-        report = request.config.cache.get('REPORT', None)
+        report = request.config.cache.get("REPORT", None)
         assert report is not None
 
         # Cache returns as dict
@@ -36,14 +36,10 @@ class TestReportImages:
 
         assert isinstance(images, list)
         assert len(images) > 0
-        assert all(i.startswith(b'<svg ') for i in images)
-
-    # def test_get_section_image(self):
-    #     pytest.skip('implement')
-    #     assert False
+        assert all(i.startswith(b"<svg ") for i in images)
 
     def test_get_elements(self, request):
-        report = request.config.cache.get('REPORT', None)
+        report = request.config.cache.get("REPORT", None)
         assert report is not None
 
         # Cache returns as dict
@@ -53,12 +49,12 @@ class TestReportImages:
         assert isinstance(elements, list)
         assert len(elements) > 0
 
-        graph = [e for e in elements if e.type == 'Graph'][0]
-        request.config.cache.set('GRAPH', graph)
+        graph = [e for e in elements if e.type == "Graph"][0]
+        request.config.cache.set("GRAPH", graph)
 
     def test_get_single_element(self, request):
-        report = request.config.cache.get('REPORT', None)
-        graph = request.config.cache.get('GRAPH', None)
+        report = request.config.cache.get("REPORT", None)
+        graph = request.config.cache.get("GRAPH", None)
         assert report is not None
         assert graph is not None
 
@@ -67,7 +63,7 @@ class TestReportImages:
         graph = RestObj(graph)
 
         # Get image with default size, specified as string
-        images = report_images.get_images(report, size='640x480', elements=graph)
+        images = report_images.get_images(report, size="640x480", elements=graph)
         assert isinstance(images, list)
         assert len(images) == 1
 
