@@ -14,17 +14,17 @@ from sasctl.utils.pyml2ds import pyml2ds
 
 
 dirname = os.path.dirname
-DATA_PATH = os.path.join(dirname(dirname(__file__)), 'pyml2ds_data')
+DATA_PATH = os.path.join(dirname(dirname(__file__)), "pyml2ds_data")
 
 
 @pytest.mark.skip(
-    'Pickle no longer loads with latest version of sklearn.  Rework to build model instead of loading.'
+    "Pickle no longer loads with latest version of sklearn.  Rework to build model instead of loading."
 )
 def test_xgb2ds():
-    pytest.importorskip('xgboost')
+    pytest.importorskip("xgboost")
 
-    IN_PKL = os.path.join(DATA_PATH, 'xgb.pkl')
-    EXPECTED_SAS = os.path.join(DATA_PATH, 'xgb_datastep')
+    IN_PKL = os.path.join(DATA_PATH, "xgb.pkl")
+    EXPECTED_SAS = os.path.join(DATA_PATH, "xgb_datastep")
 
     from sasctl.utils.pyml2ds.connectors.ensembles.xgb import XgbTreeParser
 
@@ -42,22 +42,22 @@ def test_xgb2ds():
     test_parser = TestXgbTreeParser()
 
     with mock.patch(
-        'sasctl.utils.pyml2ds.connectors.ensembles.xgb.XgbTreeParser'
+        "sasctl.utils.pyml2ds.connectors.ensembles.xgb.XgbTreeParser"
     ) as parser:
         parser.return_value = test_parser
         result = pyml2ds(IN_PKL)
 
-    with open(EXPECTED_SAS, 'r') as f:
+    with open(EXPECTED_SAS, "r") as f:
         expected = f.read()
 
     assert result == expected
 
 
 def test_lgb2ds():
-    pytest.importorskip('lightgbm')
+    pytest.importorskip("lightgbm")
 
-    IN_PKL = os.path.join(DATA_PATH, 'lgb.pkl')
-    EXPECTED_SAS = os.path.join(DATA_PATH, 'lgb_datastep')
+    IN_PKL = os.path.join(DATA_PATH, "lgb.pkl")
+    EXPECTED_SAS = os.path.join(DATA_PATH, "lgb_datastep")
 
     from sasctl.utils.pyml2ds.connectors.ensembles.lgb import LightgbmTreeParser
 
@@ -75,19 +75,19 @@ def test_lgb2ds():
     test_parser = TestLightgbmTreeParser()
 
     with mock.patch(
-        'sasctl.utils.pyml2ds.connectors.ensembles.lgb.LightgbmTreeParser'
+        "sasctl.utils.pyml2ds.connectors.ensembles.lgb.LightgbmTreeParser"
     ) as parser:
         parser.return_value = test_parser
         result = pyml2ds(IN_PKL)
 
-    with open(EXPECTED_SAS, 'r') as f:
+    with open(EXPECTED_SAS, "r") as f:
         expected = f.read()
     assert result == expected
 
 
 def test_gbm2ds():
-    IN_PKL = os.path.join(DATA_PATH, 'gbm.pmml')
-    EXPECTED_SAS = os.path.join(DATA_PATH, 'gbm_datastep')
+    IN_PKL = os.path.join(DATA_PATH, "gbm.pmml")
+    EXPECTED_SAS = os.path.join(DATA_PATH, "gbm_datastep")
 
     from sasctl.utils.pyml2ds.connectors.ensembles.pmml import PmmlTreeParser
 
@@ -105,12 +105,12 @@ def test_gbm2ds():
     test_parser = TestPmmlTreeParser()
 
     with mock.patch(
-        'sasctl.utils.pyml2ds.connectors.ensembles.pmml.PmmlTreeParser'
+        "sasctl.utils.pyml2ds.connectors.ensembles.pmml.PmmlTreeParser"
     ) as parser:
         parser.return_value = test_parser
         result = pyml2ds(IN_PKL)
 
-    with open(EXPECTED_SAS, 'r') as f:
+    with open(EXPECTED_SAS, "r") as f:
         expected = f.read()
     assert result == expected
 
@@ -121,17 +121,17 @@ def test_path_input(tmpdir_factory):
     from sasctl.utils.pyml2ds import pyml2ds
 
     # The target "model" to use
-    target = {'msg': 'hello world'}
+    target = {"msg": "hello world"}
 
     # Pickle the "model" to a file
-    temp_dir = tmpdir_factory.mktemp('pyml2ds')
-    in_file = str(temp_dir.join('model.pkl'))
-    out_file = str(temp_dir.join('model.sas'))
-    with open(in_file, 'wb') as f:
+    temp_dir = tmpdir_factory.mktemp("pyml2ds")
+    in_file = str(temp_dir.join("model.pkl"))
+    out_file = str(temp_dir.join("model.sas"))
+    with open(in_file, "wb") as f:
         pickle.dump(target, f)
 
-    with mock.patch('sasctl.utils.pyml2ds.core._check_type') as check:
-        check.translate.return_value = 'translated'
+    with mock.patch("sasctl.utils.pyml2ds.core._check_type") as check:
+        check.translate.return_value = "translated"
         pyml2ds(in_file, out_file)
 
     # Verify _check_type should have been called with the "model"
@@ -146,14 +146,14 @@ def test_file_input():
     from sasctl.utils.pyml2ds import pyml2ds
 
     # The target "model" to use
-    target = {'msg': 'hello world'}
+    target = {"msg": "hello world"}
 
     # Pickle the "model" to a file-like object
     in_file = io.BytesIO(pickle.dumps(target))
-    out_file = 'model.sas'
+    out_file = "model.sas"
 
-    with mock.patch('sasctl.utils.pyml2ds.core._check_type') as check:
-        check.translate.return_value = 'translated'
+    with mock.patch("sasctl.utils.pyml2ds.core._check_type") as check:
+        check.translate.return_value = "translated"
         pyml2ds(in_file, out_file)
 
     # Verify _check_type should have been called with the "model"
@@ -167,14 +167,14 @@ def test_pickle_input():
     from sasctl.utils.pyml2ds import pyml2ds
 
     # The target "model" to use
-    target = {'msg': 'hello world'}
+    target = {"msg": "hello world"}
 
     # Pickle the "model" to a file-like object
     in_file = pickle.dumps(target)
-    out_file = 'model.sas'
+    out_file = "model.sas"
 
-    with mock.patch('sasctl.utils.pyml2ds.core._check_type') as check:
-        check.translate.return_value = 'translated'
+    with mock.patch("sasctl.utils.pyml2ds.core._check_type") as check:
+        check.translate.return_value = "translated"
         pyml2ds(in_file, out_file)
 
     # Verify _check_type should have been called with the "model"
