@@ -13,22 +13,22 @@ def test_restobj():
     o = RestObj(a=1, b=2)
 
     assert o.a == 1
-    assert o['a'] == 1
+    assert o["a"] == 1
     assert o.b == 2
-    assert o['b'] == 2
+    assert o["b"] == 2
 
     with pytest.raises(AttributeError) as e:
         print(o.missingattribute)
     assert "missingattribute" in str(e.value)
 
     with pytest.raises(KeyError):
-        print(o['c'])
+        print(o["c"])
 
-    setattr(o, 'c', 'attribute')
-    assert o.c == 'attribute'
+    setattr(o, "c", "attribute")
+    assert o.c == "attribute"
 
     with pytest.raises(KeyError):
-        print(o['c'])
+        print(o["c"])
 
 
 def test_repr():
@@ -39,15 +39,15 @@ def test_repr():
 
 
 def test_str():
-    assert str(RestObj(name='test', id=1)) == 'test'
-    assert str(RestObj(id=1)) == '1'
-    assert str({'var': 'test'}) in str(RestObj(var='test'))
+    assert str(RestObj(name="test", id=1)) == "test"
+    assert str(RestObj(id=1)) == "1"
+    assert str({"var": "test"}) in str(RestObj(var="test"))
 
 
 def test_pickle():
     import pickle
 
-    obj = RestObj(name='test', id=1)
+    obj = RestObj(name="test", id=1)
 
     pickled = pickle.dumps(obj)
     new_obj = pickle.loads(pickled)
